added CSWAP2 opcode to processor
